+++ sample_building/sample_building.py
@@ -5,7 +5,7 @@
 # Functions to process and build samples,
 # including getting phase shifts from beam directions,
 # and functions for creating samples
-# as well as plotting them and shifting them
+# as well as shifting them as required.
 
 from scipy.fftpack import fft
 from scipy.constants import speed_of_light
@@ -14,7 +14,6 @@ import numpy as np
 import math
 import json
 from datetime import datetime
-import matplotlib.pyplot as plt
 from experiment_prototype.experiment_exception import ExperimentException
 
 def resolve_imaging_directions(beamdirs_list, num_antennas, antenna_spacing):
@@ -251,68 +250,6 @@ def shift_samples(basic_samples, phshift, amplitude):
     return samples
 
 
-def plot_samples(filename, samplesa, **kwargs):
-    """
-    Plot samples to a file for testing.
-    
-    :param filename: the filename to save the plot as.
-    :param samplesa: Some samples to plot
-    :param kwargs: any more sample arrays to plot.
-    """
-
-    more_samples = dict(kwargs)
-
-    fig, smpplot = plt.subplots(1, 1)
-    smpplot.plot(range(0, samplesa.shape[0]), samplesa)
-    for sample_name, sample_array in more_samples.items():
-        smpplot.plot(range(0, sample_array.shape[0]), sample_array)
-    # plt.ylim([-1, 1])
-    # plt.xlim([0, len(samplesa)])
-    fig.savefig(filename)
-    plt.close(fig)
-
-
-def plot_fft(filename, samplesa, rate):
-    """
-    For plotting the fft of test samples. 
-    
-    Plot the double-sided FFT of the samples in Hz (-fs/2 to +fs/2)
-    
-    :param filename: The filename to save the plot as. 
-    :param samplesa: The time-domain samples to take the fft of.
-    :param rate: The sampling rate that the samples were taken at (Hz).
-    """
-
-    fft_samps = fft(samplesa)
-    T = 1.0 / float(rate)
-    num_samps = len(samplesa)
-    xf = np.linspace(-1.0 / (2.0 * T), 1.0 / (2.0 * T), num_samps)
-    # print len(xf), len(fft_samps)
-    fig, smpplt = plt.subplots(1, 1)
-    fft_to_plot = np.empty([num_samps], dtype=np.complex64)
-    if num_samps % 2 == 1:
-        halfway = (num_samps + 1) / 2
-        for sample in range(halfway, num_samps):
-            fft_to_plot[sample - halfway] = fft_samps[sample]
-            # Move negative samples to start for plot
-        for sample in range(0, halfway):
-            fft_to_plot[sample + halfway - 1] = fft_samps[sample]
-            # Move positive samples at end
-    else:
-        halfway = num_samps / 2
-        for sample in range(halfway, num_samps):
-            fft_to_plot[sample - halfway] = fft_samps[sample]
-            # Move negative samples to start for plot
-        for sample in range(0, halfway):
-            fft_to_plot[sample + halfway] = fft_samps[sample]
-            # Move positive samples at end
-    smpplt.plot(xf, 1.0 / num_samps * np.abs(fft_to_plot))
-    #    plt.xlim([-2500000,-2000000])
-    fig.savefig(filename)
-    plt.close(fig)
-    return None
-
-
 def make_pulse_samples(pulse_list, power_divider, exp_slices, slice_to_beamdir_dict, txctrfreq,
                        txrate, options):
     """
@@ -365,7 +302,7 @@ def make_pulse_samples(pulse_list, power_divider, exp_slices, slice_to_beamdir_d
     # complex arrays (one for each possible tx antenna).
 
     #print type(pulse_list[0]), type(pulse_list[0]['samples']), type(pulse_list[0]['samples'][0])
-    plot_samples("samples.png", pulse_list[0]['samples'][0])
+    #plot_samples("samples.png", pulse_list[0]['samples'][0])
 
     # determine how long the combined pulse will be in number of samples, and add the key
     # 'sample_number_start' for all pulses in the pulse_list.
